pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpha-seeker"
version = "0.1.0"
description = "Multi-Agent LangGraph system for Alpha Seeker - discovering alpha indicators through automated research"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Alpha Seeker Team"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.1.20",
    "langchain>=0.2.14",
    "langchain-community>=0.3.1",
    "langchain-core>=0.3.1",
    "langchain-google-community>=1.0.0",
    "langgraph-cli[inmem]>=0.3.8",
    "typing_extensions>=4.9",
    "pandas>=2.1",
    "numpy>=1.25",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "typer>=0.12",
    "requests>=2.28",
    "lxml>=4.9", # for pandas.read_html
    "pydantic>=2.0",
    "asyncio-throttle>=1.0",
    "duckduckgo-search>=8.0",
    "langchain-google-genai>=2.1.9",
    "pillow>=11.3.0",
    "sentinelhub>=3.10.0",
    "streamlit>=1.48.1",
    "plotly>=6.3.0",
    "altair>=5.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
alpha-seeker = "alpha_seeker.main:app"

[project.urls]
Homepage = "https://github.com/tilaksharma/alpha-seeker"
Repository = "https://github.com/tilaksharma/alpha-seeker"
Issues = "https://github.com/tilaksharma/alpha-seeker/issues"

[tool.setuptools]
packages = ["alpha_seeker"]

[tool.setuptools.package-data]
alpha_seeker = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["alpha_seeker"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langchain.*",
    "pandas.*",
    "numpy.*",
]
ignore_missing_imports = true
